requirements.txt
chromadb
langchain-ollama
python-dotenv
langchain_community
langchain_text_splitters
pypdf
langchain-huggingface
langchain_chroma
unstructured
python-pptx
fastapi[all]
jinja2